core(remove universe level of type for now)
This is because later we would like to dig into a stable sized type
theory, this will make huge impact to this part, and we also haven't
had chance to write wrong code that runs into universe problem. The
first level mission now is fixing the implicit with complete solution.

Signed-off-by: Lîm Tsú-thuàn <dannypsnl@gmail.com>